Moved `MessageParser` into common directory
- Updated relevant imports to use the new path
