+++ code/common/command_management/invoked_command_handler.py
@@ -2,12 +2,12 @@ import asyncio
 import logging
 from typing import Callable
 
-from core.message_parser import MessageParser
 from common.configuration import Configuration
 from common.command_management.command_reconstructor import CommandReconstructor
 from common.command_management.invoked_command import InvokedCommand
 from common.database.database_manager import DatabaseManager
 from common.logging import Logging
+from common.message_parser import MessageParser
 from common.module.module import Module
 
 from discord import Interaction, Member
